pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gametranslator"
version = "0.1.0"
description = "A tool for translating game text resources with floating window"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "GameTranslator Team"}
]
dependencies = [
    # UI框架
    # 屏幕捕获和OCR
    "pytesseract>=0.3.10",
    "pyautogui>=0.9.53",
    "opencv-python>=4.7.0",
    "pillow>=9.4.0",
    # 翻译API
    "requests>=2.28.2",
    # 数据存储
    "sqlalchemy>=2.0.0",
    # 工具库
    "pyyaml>=6.0",
    "click>=8.1.3",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "pynput>=1.8.1",
    "pyside6>=6.9.2",
    "pywin32>=300",
    "chromadb>=1.1.0",
    "customtkinter>=5.2.2",
    "openai>=2.3.0",
    "pytest",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "black>=23.3.0",
    "isort>=5.12.0",
    "mypy>=1.3.0",
]

[project.scripts]
gametranslator = "gametranslator.main:main"

[tool.setuptools]
package-dir = {"" = "src"}
